requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-genai==0.5.0
mcp==1.0.0
fastmcp==0.1.0
httpx==0.25.2
python-dotenv==1.0.0
